220501_충돌처리 수정

+++ 2_Team/2_Team/Thorn.cpp
@@ -39,8 +39,8 @@ void CThorn::Render(HDC _hDC)
 {
 	int		iScrollX = (int)SCROLLMGR->Get_ScrollX();
 
-	MoveToEx(_hDC, m_tInfo.fX - 50 + iScrollX, m_tInfo.fY - 300, nullptr);
-	LineTo(_hDC, m_tInfo.fX - 50 + iScrollX, m_tInfo.fY + 300);
+	MoveToEx(_hDC, LONG(m_tInfo.fX - 50 + iScrollX), LONG(m_tInfo.fY - 300), nullptr);
+	LineTo(_hDC, LONG(m_tInfo.fX - 50 + iScrollX), LONG(m_tInfo.fY + 300));
 	
 	MoveToEx(_hDC, m_tInfo.fX - 50+ iScrollX, m_tInfo.fY - 300, nullptr);
 	LineTo(_hDC, m_tInfo.fX + 50+ iScrollX, m_tInfo.fY - 300);
@@ -49,15 +49,15 @@ void CThorn::Render(HDC _hDC)
 	//LineTo(_hDC, m_tInfo.fX + iScrollX, m_tInfo.fY - 300 + 30);
 	for (int i = 0; i < 10; ++i)
 	{
-		MoveToEx(_hDC, m_tInfo.fX + 50 + iScrollX, m_tInfo.fY - 300 + i*60, nullptr);
-		LineTo(_hDC, m_tInfo.fX + iScrollX, m_tInfo.fY - 300 + 30 + i*60);
+		MoveToEx(_hDC, LONG(m_tInfo.fX + 50 + iScrollX), LONG(m_tInfo.fY - 300 + i*60), nullptr);
+		LineTo(_hDC, LONG(m_tInfo.fX + iScrollX), LONG(m_tInfo.fY - 300 + 30 + i*60));
 
 		MoveToEx(_hDC, m_tInfo.fX + iScrollX, m_tInfo.fY - 300 + 30 + i * 60, nullptr);
 		LineTo(_hDC, m_tInfo.fX + 50 + iScrollX, m_tInfo.fY - 300 + 60 + i * 60);
 	}
 
-	MoveToEx(_hDC, m_tInfo.fX - 50+ iScrollX, m_tInfo.fY + 300, nullptr);
-	LineTo(_hDC, m_tInfo.fX + 50+ iScrollX, m_tInfo.fY + 300);
+	MoveToEx(_hDC, LONG(m_tInfo.fX - 50+ iScrollX), LONG(m_tInfo.fY + 300), nullptr);
+	LineTo(_hDC, LONG(m_tInfo.fX + 50+ iScrollX), LONG(m_tInfo.fY + 300));
 
 
 }
